requirements.txt
arrow==1.2.3
asgiref==3.7.2
attrs==23.1.0
binaryornot==0.4.4
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
Django==4.2.4
django-cors-headers==4.2.0
django-environ==0.10.0
django-modeltranslation==0.18.11
djangorestframework==3.14.0
drf-spectacular==0.26.4
exceptiongroup==1.1.2
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
packaging==23.1
Pillow==10.0.0
pluggy==1.2.0
psycopg2-binary==2.9.6
Pygments==2.16.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
whitenoise==6.5.0
